build properly on windows
